+++ packages/aws-cdk-local-lambda/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./extract/index.js";
+export * from "./server/index.js";
+export * from "./types.js";