+++ packages/cdk-lambda-local/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./extract/index";
-export * from "./server/index";
-export * from "./types";